requirements.txt
python-docx==1.1.0
Pillow==10.2.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
